Add an additional epoch-based mechanism to avoid running passes on function if they will not detect anything new.
The idea is that some passes are purely local and thus do not depend on any other functions or analysis (e.g. SILCombine, SimplifyCFG, etc).
Epochs are associated with SIL functions. Every time an instruction is added to or removed from a function, its epoch is increased.
After run of a given pass, the pass manager remembers the epoch of a function. When it is about to run the same pass on the same function next time, it checks if the epoch has changed. If it is unchanged, then running the pass will not produce any new results.

TODO:
- Passes should have an API to check if they are epoch-based or not.
- If an analysis is invalidated, then a pass depending on this analysis should be invalidated/rerun even if the function itself has not changed.