And almost working version of the NonAtomicRC analysis.
It is now more or less feature complete. It performs a proper forward data flow analysis. It is then eliminated redundant make_mutable calls. It also marks RC operations inside the uniqueness region as thread-local.

Still needs to be tested for correctness as there may be some corner cases which are not covered.
And tests still need to be written.