Re-factored the code.
It does not use the auto-linking trick anymore and generates wrappers in IRGen.
It does not use the new calling convention for objc_retain/objc_release or anything escaping into the Obj-C runtime.
It supports the generation of wrappers even when the new calling convention is not used (via Config.h)