Add a new array semantics called "array.gurantee_mutable"
Semantics "array.gurantee_mutable" means that the array is unique after this call.
It essentially means that the function annotated with this semantics internally performs makeUnique and may be does something else.
Therefore, invocations of such a function cannot be removed even if it is known that the array is unique.