Add support for using non-atomic versions of functions.
There is a new semantics called "generate.nonatomic".
If a function is annotated with such a semantics attribute, it means that this function has a non-atomic version, where the Self parameter is known to be unique and thus thread-local. Therefore, non-atomic RC operations can be used with it.

The non-atomic version of the function annotated with such an attribute is generated automatically by the compiler. This is done using the following rules:
1) In general, the newly generated function is a clone of the original function
2) All RC operations on the Self argument (or its projections) are replaced by non-atomic RC operations
3) All calls of functions annotated by @_semantics("generate.nonatomic") are replaced by calls of non-atomic versions of these functions, which are generated on demand if necessary.
4) All is_unque instructions on the Self arguments are folded into a success outcome.