stdlib: Add some @_versioned declarations to fix resilient build
A recent change made accessibility checking stricter. This had some
fallout on the half-baked @_versioned attribute, where we could no
longer define @_versioned members on a non-@_versioned type.

This was wrong anyway (and will be diagnosed when we add proper
diagnostics for @_versioned), because type metadata for the
internal type did not get the right linkage, but it used to work
as long as you didn't try to get the type metadata at runtime.

This patch adds @_versioned attributes to the right types now that
this broken behavior is gone.

As a result, _Variant{Set,Dictionary}Storage becomes resilient
(non-@_versioned internal types are not resilient), which broke
the uniqueness check, where we assumed that the 'native' case had
all spare bits set to zero; to make it work again, this uses
the new @_fragile_case_order attribute.

Fixes <rdar://problem/28409189>.