IRGen: Add @_fragile_case_order attribute
This allows you to define a resilient enum while maintaining full
control over the order of cases, which are numbered as they are
written in the source.

This means in particular that for a multi-payload enum E where all
cases are pointers, the first case of a @_fragile_case_order
enum has all spare bits as zero, allowing you to bitwise cast
E to the type of the first payload and get a valid pointer.

This is a very specific feature to allow making the
_Variant{Set,Dictionary}Storage types in the standard library
resilient. There might be a better way of doing this, or we
might just decide to make these @_fixed_layout in the future.