Fix the original version of the partial specialization implementation.
This approach does not try to create a new generic signature with  a different number of generic type parameters. Instead, it simply builds a new signature based on the old one and adds same type requirements for those generic type parameters that are concrete according to the partial substitution. This way, the signature has exactly the same generic parameter types, just with more requirements. It is much easier to create than using the other approach, because it does not require any complex re-mappings of generic types and archetypes.