Initial version of the alternative partial specialization implementation.
Overall idea:
Create a new generic signature based on the generic signature of the callee and a set of apply substitutions. The new signature should contain generic parameters for all the caller's archetypes used in the apply's substitutions. It should also have all the requirements derived from the generic signature of the callee.

Status: It is not working yet due to some crashes during stdlib builds.