Initial implementation of the @_specialize attribute supporting layout constraints and partial specializations.
One can use the `@_specialize` attribute to force creation of specializations of a given function for a given set of generic parameters.

The new `@_specialize` attribute has a syntax like this:

```swift
@_specialize(exported: false, kind: full, where K == Int, V == Int)
@_specialize(exported: false, kind: partial, where K: _Trivial64)
func dictFunction<K, V>(dict: Dictionary<K, V>) {
...
}
```

If `exported` is set, the corresponding specialization would have a public visibility and can be used by clients.
If `exported` is omitted, it's value is assumed to be `false`.

If `kind` is `full` it means a full specialization.
If `kind` is `partial` it means a partial specialization.
If `kind` is omitted, its value is assumed to be `full`.

One can use in the `where` clause of the `@_specialize` attribute the newly introduced layout constraints like e.g. _Trivial, _Trivial8, _Trivial16, _Trivial32, _Trivial64 or _RefCountedObject.

These layout constraints have the following semantics:
- `_TrivialN` means an value of a trivial type, which is N bits long. Trivial means that it can be copied by means of a bitwise copy and does not contain any references. If `N` is not specified then object can be of any size, but should be still trivial.
- `_RefCountedObject ` means a reference counted object, e.g. it can be a reference to an instance of a class.