A lot of fixes towards more workable solution.
It now correctly lazily emits in IRGen all the required metadata and functions for structs, classes and protocols.
The only unresolved symbols are those Swift functions that are referenced by the runtime by name.
I need to emit some empty stubs for them, which would fail with an error message at runtime.