Add a new function specialization optimization pass
This pass specializes functions for constant argument values and replaces the calls by calls of a specialized function.
At the moment, it only performs this optimization for the constant arguments which are function references and only if the same effect cannot be achieved by inlining the original function (e.g. when the callee is self-recursive, contains dynamic self, etc).

This pass does not apply to many functions in the standard library, but when it does it produces significant speedups. E.g. sorting functions are 3x-4x faster with this optimization.

Fixes rdar://31939398