Do not link against any external non-public symbols
It was possible to link against an external symbol, which is was not public at the SIL level. This was done to reduce the client code size, but it is not valid in a general case. Only public external symbols should be referenced. In all other cases the compiler should emit its own copy of the symbol.