Make `@_versioned` + `@inline(never)` a replacement for `@_semantics("stdlib_binary_only")` in the -sil-serialize-all mode
`@_semantics("stdlib_binary_only")` was a hack to make sure that SIL bodies of some functions are not imported into client code and are only present as public symbols in the object file of the module that defines them.

Replace this hack with a more future-proof solution, which is compatible with the upcoming resilience features. You can now use a combination of `@_versioned` and `@inline(never)` to achieve the same effect.

In sil-serialize-all mode, using these annotations would result in:
- SIL function body not being serialized
- The annotated function itself will be IRGened and made available as a public symbol in the object file of the module that defines it.